backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic[email]==2.9.2
httpx>=0.26,<0.28
python-multipart==0.0.6
bcrypt==4.1.2
slowapi==0.1.9
supabase==2.9.0

# Optional integrations
mercadopago==2.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3

# Monitoring
sentry-sdk[fastapi]==2.21.0

# Backup
boto3==1.35.94
